WIP 2023-01-24
 - some more C++ std updates
 - add script for building docs
 - write a test framework and some initial tests
 - add GN template for tests
 - add script for running tests
 - some C++ lib changes to support building on the host for testing
 - fix some bugs revealed by the tests!
 - fix some build warnings

+++ System/ABI/Dwarf/Source/System/ABI/Dwarf/FDE.cc
@@ -452,7 +452,7 @@ DwarfFDE DwarfFDE::DecodeFrom(const std::byte*& ptr, std::uintptr_t text_base, s
     fde.m_codeLength = instructionBytes;
     fde.m_lsda = lsda;
     fde.m_instructions = ptr;
-    fde.m_instructionsLength = (fdeStart + fdeLength) - ptr;
+    fde.m_instructionsLength = static_cast<std::size_t>((fdeStart + fdeLength) - ptr);
     ptr = fdeStart + fdeLength;
     return fde;
 }
